requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
certifi==2024.2.2
charset-normalizer==3.3.2
comtypes==1.3.1
idna==3.6
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pillow==10.2.0
pypiwin32==223
pyttsx3==2.90
pywin32==306
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
Flask~=3.0.2
python-dotenv~=1.0.1